+++ src/Symfony/Bridge/Twig/Tests/Node/SearchAndRenderBlockNodeTest.php
@@ -264,10 +264,6 @@ class SearchAndRenderBlockNodeTest extends TestCase
 
     protected function getVariableGetter($name)
     {
-        if (PHP_VERSION_ID >= 70000) {
-            return sprintf('($context["%s"] ?? null)', $name, $name);
-        }
-
-        return sprintf('(isset($context["%s"]) ? $context["%s"] : null)', $name, $name);
+        return sprintf('($context["%s"] ?? null)', $name, $name);
     }
 }