+++ src/Symfony/Bridge/Twig/Tests/Node/TransNodeTest.php
@@ -40,11 +40,7 @@ class TransNodeTest extends TestCase
 
     protected function getVariableGetterWithoutStrictCheck($name)
     {
-        if (PHP_VERSION_ID >= 70000) {
-            return sprintf('($context["%s"] ?? null)', $name, $name);
-        }
-
-        return sprintf('(isset($context["%s"]) ? $context["%s"] : null)', $name, $name);
+        return sprintf('($context["%s"] ?? null)', $name, $name);
     }
 
     protected function getVariableGetterWithStrictCheck($name)
@@ -53,10 +49,6 @@ class TransNodeTest extends TestCase
             return sprintf('(isset($context["%s"]) || array_key_exists("%s", $context) ? $context["%s"] : (function () { throw new Twig_Error_Runtime(\'Variable "%s" does not exist.\', 0, $this->getSourceContext()); })())', $name, $name, $name, $name);
         }
 
-        if (PHP_VERSION_ID >= 70000) {
-            return sprintf('($context["%s"] ?? $this->getContext($context, "%s"))', $name, $name, $name);
-        }
-
-        return sprintf('(isset($context["%s"]) ? $context["%s"] : $this->getContext($context, "%s"))', $name, $name, $name);
+        return sprintf('($context["%s"] ?? $this->getContext($context, "%s"))', $name, $name, $name);
     }
 }