+++ src/Symfony/Component/DependencyInjection/Tests/Loader/IniFileLoaderTest.php
@@ -45,7 +45,6 @@ class IniFileLoaderTest extends TestCase
 
     /**
      * @dataProvider getTypeConversions
-     * @requires PHP 5.6.1
      * This test illustrates where our conversions differs from INI_SCANNER_TYPED introduced in PHP 5.6.1
      */
     public function testTypeConversionsWithNativePhp($key, $value, $supported)